Undo refactoring related changes.
