+++ operator/src/main/java/oracle/kubernetes/operator/WaitForReadyStep.java
@@ -4,17 +4,25 @@
 package oracle.kubernetes.operator;
 
 import java.util.Optional;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Consumer;
 
+import io.kubernetes.client.openapi.models.V1Job;
 import io.kubernetes.client.openapi.models.V1ObjectMeta;
+import io.kubernetes.client.openapi.models.V1Pod;
+import oracle.kubernetes.operator.calls.CallResponse;
+import oracle.kubernetes.operator.helpers.CallBuilder;
 import oracle.kubernetes.operator.helpers.DomainPresenceInfo;
 import oracle.kubernetes.operator.helpers.ResponseStep;
+import oracle.kubernetes.operator.steps.DefaultResponseStep;
 import oracle.kubernetes.operator.work.AsyncFiber;
 import oracle.kubernetes.operator.work.NextAction;
 import oracle.kubernetes.operator.work.Packet;
 import oracle.kubernetes.operator.work.Step;
+import oracle.kubernetes.weblogic.domain.model.Domain;
 
+import static oracle.kubernetes.operator.ProcessingConstants.MAKE_RIGHT_DOMAIN_OPERATION;
 import static oracle.kubernetes.operator.helpers.KubernetesUtils.getDomainUidLabel;
 
 /**
@@ -23,7 +31,6 @@ import static oracle.kubernetes.operator.helpers.KubernetesUtils.getDomainUidLab
  * @param <T> the type of resource handled by this step
  */
 abstract class WaitForReadyStep<T> extends Step {
-
   private static final int DEFAULT_RECHECK_SECONDS = 5;
   private static final int DEFAULT_RECHECK_COUNT = 60;
 
@@ -178,15 +185,13 @@ abstract class WaitForReadyStep<T> extends Step {
             null);
   }
 
-  Step createReadAndIfReadyCheckStep(Callback callback) {
+  private Step createReadAndIfReadyCheckStep(Callback callback) {
     if (initialResource != null) {
       return createReadAsyncStep(getName(), getNamespace(), getDomainUid(), resumeIfReady(callback));
     } else {
-      return new ReadAndIfReadyCheckStep(getName(), callback, resumeIfReady(callback), getNext());
+      return new ReadAndIfReadyCheckStep(getName(), callback, getNext());
     }
   }
-
-  protected abstract ResponseStep resumeIfReady(Callback callback);
 
   private String getNamespace() {
     return getMetadata(initialResource).getNamespace();
@@ -200,28 +205,82 @@ abstract class WaitForReadyStep<T> extends Step {
     return initialResource != null ? getMetadata(initialResource).getName() : resourceName;
   }
 
+  private DefaultResponseStep<T> resumeIfReady(Callback callback) {
+    return new DefaultResponseStep<>(null) {
+      @Override
+      public NextAction onSuccess(Packet packet, CallResponse<T> callResponse) {
+        DomainPresenceInfo info = packet.getSpi(DomainPresenceInfo.class);
+        if ((callResponse != null) && (callResponse.getResult() instanceof V1Pod)) {
+          V1Pod pod = (V1Pod) callResponse.getResult();
+          Optional.ofNullable(info)
+                  .ifPresent(i -> i.setServerPodFromEvent(getPodLabel(pod, LabelConstants.SERVERNAME_LABEL), pod));
+        }
+        if (isReady(callResponse.getResult())) {
+          callback.proceedFromWait(callResponse.getResult());
+          Optional.ofNullable(info).ifPresent(i -> i.resetWatchBackstopRecheckCount());
+          return doNext(packet);
+        }
+        if (shouldWait(callResponse, info)) {
+          // Watch backstop recheck count is less than or equal to the configured recheck count, delay.
+          return doDelay(createReadAndIfReadyCheckStep(callback), packet,
+                  getWatchBackstopRecheckDelaySeconds(), TimeUnit.SECONDS);
+        } else {
+          // Watch backstop recheck count is more than configured recheck count, proceed to make-right step.
+          return doNext(new CallBuilder().readDomainAsync(info.getDomainUid(),
+                  info.getNamespace(), new MakeRightDomainStep(null)), packet);
+        }
+      }
+
+      private boolean shouldWait(CallResponse<T> callResponse, DomainPresenceInfo info) {
+        return info == null || (info.incrementAndGetWatchBackstopRecheckCount() <= getWatchBackstopRecheckCount())
+                || ((callResponse != null) && (callResponse.getResult() instanceof V1Job));
+      }
+
+      private String getPodLabel(V1Pod pod, String labelName) {
+        return Optional.ofNullable(pod)
+                .map(V1Pod::getMetadata)
+                .map(V1ObjectMeta::getLabels)
+                .map(m -> m.get(labelName))
+                .orElse(null);
+      }
+    };
+  }
 
   private class ReadAndIfReadyCheckStep extends Step {
     private final Callback callback;
     private final String resourceName;
-    private final ResponseStep responseStep;
 
-    ReadAndIfReadyCheckStep(String resourceName, Callback callback, ResponseStep responseStep, Step next) {
+    ReadAndIfReadyCheckStep(String resourceName, Callback callback, Step next) {
       super(next);
       this.callback = callback;
       this.resourceName = resourceName;
-      this.responseStep = responseStep;
     }
 
     @Override
     public NextAction apply(Packet packet) {
       DomainPresenceInfo info = packet.getSpi(DomainPresenceInfo.class);
       return doNext(createReadAsyncStep(resourceName, info.getNamespace(),
-              info.getDomainUid(), responseStep), packet);
+              info.getDomainUid(), resumeIfReady(callback)), packet);
+    }
+
+  }
+
+  private class MakeRightDomainStep extends DefaultResponseStep {
+    MakeRightDomainStep(Step next) {
+      super(next);
+    }
+
+    @Override
+    public NextAction onSuccess(Packet packet, CallResponse callResponse) {
+      MakeRightDomainOperation makeRightDomainOperation =
+              (MakeRightDomainOperation)packet.get(MAKE_RIGHT_DOMAIN_OPERATION);
+      makeRightDomainOperation.setLiveInfo(new DomainPresenceInfo((Domain)callResponse.getResult()));
+      makeRightDomainOperation.withExplicitRecheck().interrupt().execute();
+      return super.onSuccess(packet, callResponse);
     }
   }
 
-  class Callback implements Consumer<T> {
+  private class Callback implements Consumer<T> {
     private final AsyncFiber fiber;
     private final Packet packet;
     private final AtomicBoolean didResume = new AtomicBoolean(false);
@@ -240,7 +299,7 @@ abstract class WaitForReadyStep<T> extends Step {
     }
 
     // The resource has now either completed or failed, so we can continue processing.
-    void proceedFromWait(T resource) {
+    private void proceedFromWait(T resource) {
       removeCallback(getName(), this);
       if (mayResumeFiber()) {
         handleResourceReady(fiber, packet, resource);
